ch03/pyproject.toml
[project]
name = "ch03"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.28.0",
    "ipykernel>=6.30.1",
    "jupyter>=1.1.1",
    "jupyterlab>=4.4.5",
    "langchain>=0.3.27",
    "langchain-google-genai>=2.1.9",
    "python-dotenv>=1.1.1",
    "streamlit>=1.48.0",
]
